+++ src/backend/core/AbstractColumn.h
@@ -77,10 +77,10 @@ public:
 //		UInt8 = 21,	// quint8 (unsigned char)
 //		Int16 = 22,	// qint16 (short)
 //		UInt16 = 23,	// quint16 (unsigned short)
-		Integer = 24,	// qint32 (int)
-//		UInt32 = 25,	// quint32 (unsigned int)
-//		Int64 = 26,	// qint64 (long)
-//		UInt64 = 27,	// quint64 (unsigned long)
+		Integer = 24,	// qint64 (long)
+//		UInt64 = 25,	// quint64 (unsigned long)
+//		Int32 = 26,	// qint32 (int)
+//		UInt32 = 27,	// quint32 (unsigned int)
 		// MISC
 		// QBrush = 30
 		// QColor
@@ -191,9 +191,9 @@ public:
 	virtual double valueAt(int row) const;
 	virtual void setValueAt(int row, double new_value);
 	virtual void replaceValues(int first, const QVector<double>& new_values);
-	virtual int integerAt(int row) const;
-	virtual void setIntegerAt(int row, int new_value);
-	virtual void replaceInteger(int first, const QVector<int>& new_values);
+	virtual long integerAt(int row) const;
+	virtual void setIntegerAt(int row, long new_value);
+	virtual void replaceInteger(int first, const QVector<long>& new_values);
 	virtual Properties properties() const;
 
 signals:
