+++ src/backend/core/column/ColumnPrivate.h
@@ -110,9 +110,9 @@ public:
 	void setValueAt(int row, double new_value);
 	void replaceValues(int first, const QVector<double>&);
 
-	int integerAt(int row) const;
-	void setIntegerAt(int row, int new_value);
-	void replaceInteger(int first, const QVector<int>&);
+	long integerAt(int row) const;
+	void setIntegerAt(int row, long new_value);
+	void replaceInteger(int first, const QVector<long>&);
 
 	void updateProperties();
